Moved views to accounting.py
