+++ corehq/apps/domain/urls.py
@@ -17,44 +17,46 @@ from corehq.apps.callcenter.views import CallCenterOwnerOptionsView
 from corehq.apps.domain.forms import ConfidentialPasswordResetForm, HQSetPasswordForm
 from corehq.apps.domain.views import (
     ActivateTransferDomainView,
-    BillingStatementPdfView,
-    BulkStripePaymentView,
-    CardView,
-    CardsView,
     CaseSearchConfigView,
-    ConfirmBillingAccountInfoView,
-    ConfirmSelectedPlanView,
-    ConfirmSubscriptionRenewalView,
     CreateNewExchangeSnapshotView,
-    CreditsStripePaymentView,
-    CreditsWireInvoiceView,
     DeactivateTransferDomainView,
     DefaultProjectSettingsView,
-    DomainBillingStatementsView,
-    DomainSubscriptionView,
     EditBasicProjectInfoView,
-    EditExistingBillingAccountView,
     EditInternalCalculationsView,
     EditInternalDomainInfoView,
     EditMyProjectSettingsView,
     EditOpenClinicaSettingsView,
     EditPrivacySecurityView,
-    EmailOnDowngradeView,
     FlagsAndPrivilegesView,
     FeaturePreviewsView,
-    InternalSubscriptionManagementView,
-    InvoiceStripePaymentView,
     ManageProjectMediaView,
     PasswordResetView,
     RecoveryMeasuresHistory,
-    SelectPlanView,
-    SelectedEnterprisePlanView,
-    SelectedAnnualPlanView,
-    SubscriptionRenewalView,
     TransferDomainView,
-    WireInvoiceView,
     calculated_properties,
     toggle_diff,
+)
+from corehq.apps.domain.views.accounting import (
+    BillingStatementPdfView,
+    BulkStripePaymentView,
+    CardView,
+    CardsView,
+    ConfirmBillingAccountInfoView,
+    ConfirmSelectedPlanView,
+    ConfirmSubscriptionRenewalView,
+    CreditsStripePaymentView,
+    CreditsWireInvoiceView,
+    DomainBillingStatementsView,
+    DomainSubscriptionView,
+    EditExistingBillingAccountView,
+    EmailOnDowngradeView,
+    InternalSubscriptionManagementView,
+    InvoiceStripePaymentView,
+    SelectedEnterprisePlanView,
+    SelectedAnnualPlanView,
+    SelectPlanView,
+    SubscriptionRenewalView,
+    WireInvoiceView,
 )
 from corehq.apps.domain.views.base import select
 from corehq.apps.domain.views.exchange import ExchangeSnapshotsView, set_published_snapshot,
